+++ common/src/main/java/com/evandev/reliable_remover/mixin/minecraft/ServerPlayerMixin.java
@@ -40,7 +40,7 @@ public abstract class ServerPlayerMixin extends Player {
 
     @Inject(method = "attack", at = @At("HEAD"), cancellable = true)
     private void reliable_remover$attack(Entity target, CallbackInfo ci) {
-        if (RuleManager.isAttackBlocked(this.getMainHandItem(), this.level())) {
+        if (RuleManager.isAttackBlocked(this.getMainHandItem(), this.level(), target)) {
             this.displayClientMessage(Component.translatable("message.reliable_remover.attack_disabled"), true);
             ci.cancel();
         }
@@ -48,7 +48,7 @@ public abstract class ServerPlayerMixin extends Player {
 
     @Inject(method = "swing", at = @At("HEAD"), cancellable = true)
     private void reliable_remover$cancelSwing(InteractionHand hand, CallbackInfo ci) {
-        if (RuleManager.isAttackBlocked(this.getMainHandItem(), this.level())) {
+        if (RuleManager.isAttackBlocked(this.getMainHandItem(), this.level(), null)) {
             ci.cancel();
         }
     }